extra/requirements.txt
# Core Web Framework
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.7
flask-wtf==1.2.1
flask-login==0.6.3
flask-bcrypt==1.0.1
flask-moment==1.0.5
python-dotenv==1.0.1
werkzeug==3.0.1

# Background Tasks & Queue
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# AI & Computer Vision - Windows Compatible
opencv-python==4.8.1.78
face-recognition==1.3.0
# Skip dlib for now - will install separately if needed
# dlib causes build issues on Windows
Pillow>=8.3.2,<10.0.0
numpy>=1.21.0,<2.0.0
scikit-learn>=1.3.0
imageio>=2.25.0,<2.32.0
scipy>=1.9.0
matplotlib>=3.6.0

# Location & Mapping
geopy==2.4.1
folium>=0.15.0
geojson>=3.0.0

# Database Support
psycopg2-binary==2.9.9
alembic>=1.12.0

# Forms & Validation
WTForms==3.1.1
email-validator==2.1.0
validators>=0.20.0

# Security & Encryption
itsdangerous==2.1.2
markupsafe==2.1.3
bleach>=6.0.0
cryptography>=40.0.0

# File Handling & Utils
filetype>=1.2.0
requests>=2.28.0
urllib3>=1.26.0

# Date & Time
pytz>=2023.3
python-dateutil>=2.8.0

# Production Server
gunicorn>=21.0.0
gevent>=23.7.0
eventlet>=0.33.0

# Template Engine
jinja2>=3.0.0

# Fix deprecation warnings
setuptools<81